Add "isDeleted" column to tables - Menu - MenuDish - MenuPerDay - User - UserProfile
On MenuPerDayPerPerson the column was not added because we don't need to keep any previous attendance decisions.

Update queries to filter the data that is deleted except on the situation of report generation. If the admin deletes a menu that was used two months before, for example, the report for that month will still contain the data about this menu

Removed methods that were not used